# backend/booking-service/docs/swagger.yaml
basePath: /
definitions:
  internal_handler.createBookingRequest:
    properties:
      payment_id:
        type: string
      session_id:
        type: integer
      total_amount_cents:
        type: integer
    type: object
...
  internal_handler.createMovieRequest:
    properties:
      banner_url:
        example: https://example.com/banner.jpg
        type: string
      description:
        example: Epic sci-fi film
        type: string
      duration_minutes:
        example: 166
        type: integer
      poster_url:
        example: https://example.com/poster.jpg
        type: string
      rating:
        example: 8.7
        type: number
      release_date:
        example: "2024-03-01T00:00:00Z"
        type: string
      title:
        example: 'Dune: Part Two'
        type: string
      trailer_url:
        example: https://example.com/trailer.mp4
        type: string
    type: object
...
  internal_handler.updateMovieRequest:
    properties:
      banner_url:
        example: https://example.com/banner.jpg
        type: string
      description:
        example: Epic sci-fi film
        type: string
      duration_minutes:
        example: 166
        type: integer
      poster_url:
        example: https://example.com/poster.jpg
        type: string
      rating:
        example: 8.7
        type: number
      release_date:
        example: "2024-03-01T00:00:00Z"
        type: string
      title:
        example: 'Dune: Part Two'
        type: string
      trailer_url:
        example: https://example.com/trailer.mp4
        type: string
    type: object
...
  model.Booking:
    properties:
      cancelled_at:
        type: string
      created_at:
        type: string
      id:
        type: integer
      payment_id:
        type: string
      session_id:
        type: integer
      status:
        type: string
      total_amount_cents:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  model.Movie:
    properties:
      banner_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      id:
        type: integer
      poster_url:
        type: string
      rating:
        type: number
      release_date:
        type: string
      title:
        type: string
      trailer_url:
        type: string
    type: object
  model.Session:
    properties:
      available_seats:
        type: integer
      base_price_cents:
        type: integer
      cinema_address:
        type: string
      cinema_city:
        type: string
      cinema_id:
        type: integer
      cinema_name:
        type: string
      created_at:
        type: string
      end_time:
        type: string
      hall_id:
        type: integer
      hall_name:
        type: string
      id:
        type: integer
      movie_id:
        type: integer
      movie_title:
        type: string
      start_time:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: API for movie sessions and bookings
  title: Booking Service API
  version: "1.0"
paths:
...
  /admin/movies:
    get:
      description: Returns all movies for admin management
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Movie'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get admin movies
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Creates a new movie
      parameters:
      - description: Movie payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createMovieRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Movie'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create movie
      tags:
      - admin
  /admin/movies/{id}:
    delete:
      description: Deletes movie by ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete movie
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Updates movie by ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      - description: Movie payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Movie'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update movie
      tags:
      - admin
...
  /bookings:
    post:
      consumes:
      - application/json
      description: Create a new booking for the authenticated user
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: integer
      - description: Booking request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create booking
      tags:
      - bookings
  /bookings/{id}:
    get:
      description: Returns booking by booking ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get booking by ID
      tags:
      - bookings
  /bookings/{id}/cancel:
    patch:
      description: Cancels booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cancel booking
      tags:
      - bookings
  /bookings/{id}/confirm:
    patch:
      description: Confirms booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Confirm booking
      tags:
      - bookings
  /bookings/me:
    get:
      description: Returns all bookings for authenticated user
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Booking'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get my bookings
      tags:
      - bookings
  /movies:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Movie'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all movies
      tags:
      - movies
  /movies/{id}:
    get:
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Movie'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get movie by ID
      tags:
      - movies
  /movies/{id}/sessions:
    get:
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Session'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get sessions by movie ID
      tags:
      - movies
  /sessions/{id}:
    get:
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Session'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get session by ID
      tags:
      - movies
swagger: "2.0"